task/17_Global_State_Management_and_Data_Fetching: update terbaru untuk praktikum global state management - menambahkan redux dengan nama buttonSlice.js - membuat folder features serta memindahkan file todoSlice.js dan buttonSlice.js ke dalamnya - menghapus components Todo yaitu infoItem beserta file di dalamnya InfoItem.jsx dan InfoItem.module.css - memperbarui untuk components Todo yaitu InfoTodo - menambahkan redux buttonSlice.js ke dalam store.js

+++ 17_Global State Management and Data Fetching/praktikum/global-state/src/components/Todo/InfoTodo/InfoTodo.jsx
@@ -1,16 +1,15 @@
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { handleDeleteAllTodo } from "../../../store/todoSlice";
+import { handleDeleteAllTodo } from "../../../store/features/todoSlice";
+import { disabledButton } from "../../../store/features/buttonSlice";
 
-import InfoItem from "../InfoItem/InfoItem";
+import style from "./InfoTodo.module.css";
 
 const InfoTotalTodo = () => {
+	const disableButton = useSelector((state) => state.button);
 	const todos = useSelector((state) => state.todo.todos);
 	const dispatch = useDispatch();
 
-	const [disabled, setDisabled] = useState(true);
-
-	// Fitur Tambahan: Total todo yang sudah selesai
 	const getTotalCompletedTodos = () => {
 		const totalCounts = todos.reduce((total, num) => {
 			return total + num.completed;
@@ -19,7 +18,6 @@ const InfoTotalTodo = () => {
 		return totalCounts;
 	};
 
-	// Fitur Tambahan: Total todo yang belum selesai
 	const getTotalUncompletedTodos = () => {
 		const totalCounts = todos.reduce((total, num) => {
 			return total + !num.completed;
@@ -30,23 +28,27 @@ const InfoTotalTodo = () => {
 
 	useEffect(() => {
 		if (todos.length > 0) {
-			setDisabled(false);
+			dispatch(disabledButton(false));
 		} else {
-			setDisabled(true);
+			dispatch(disabledButton(true));
 		}
-	}, [todos]);
+	});
 
 	return (
 		<div>
-			<InfoItem
-				todos={todos}
-				handleDeleteAllTodo={() => {
-					dispatch(handleDeleteAllTodo(todos));
-				}}
-				totalCompletedCount={getTotalCompletedTodos()}
-				totalUncompletedCount={getTotalUncompletedTodos()}
-				disabled={disabled}
-			/>
+			<div className="d-flex justify-content-between align-items-center pt-4 pb-3">
+				<div className={`fw-semibold ${style.infoTotal}`}>{`Total Todo = ${todos.length}`}</div>
+				<div className={`fw-semibold ${style.infoTotal}`}>{`Total Completed Todo = ${getTotalCompletedTodos()}`}</div>
+				<div className={`fw-semibold ${style.infoTotal}`}>{`Total Uncompleted Todo = ${getTotalUncompletedTodos()}`}</div>
+				<button
+					onClick={() => {
+						if (window.confirm("Apakah Anda yakin ingin menghapus semua daftar pekerjaan Anda?")) dispatch(handleDeleteAllTodo(todos));
+					}}
+					className={`btn text-danger ${style.deleteAllBtn}`}
+					disabled={disableButton}>
+					Delete All Todo
+				</button>
+			</div>
 		</div>
 	);
 };
